Add AssimpTool for asset conversion and import
Introduced a new AssimpTool module with supporting classes (Converter, CameraScript, ImGUIDemo, StaticMeshDemo, and related headers) for importing, converting, and exporting 3D assets using Assimp. Added project files, resource assets, shaders, and updated solution and engine to integrate the new tool. Also included Assimp library headers and binaries, and updated related project and filter files.

+++ Main/Engine/Camera.h
@@ -21,8 +21,8 @@ public:
 	void SetNear(float value) { _near = value; }
 	void SetFar(float value) { _far = value; }
 	void SetFOV(float value) { _fov = value; }
-	void SetWidth(float value) { _width = value; }
-	void SetHeight(float value) { _height = value; }
+	void SetWidth(UINT value) { _width = value; }
+	void SetHeight(UINT value) { _height = value; }
 
 	Matrix& GetViewMatrix() { return _matView; }
 	Matrix& GetProjectionMatrix() { return _matProjection; }
@@ -34,8 +34,8 @@ private:
 	float _near		= 1.f;
 	float _far		= 1000.f;
 	float _fov		= XM_PI / 4.f;
-	float _width	= 0.f;
-	float _height	= 0.f;
+	UINT _width		= 0;
+	UINT _height	= 0;
 
 public:
 	static Matrix S_MatView;
